+++ subprojects/bounded_context/place/application/src/main/kotlin/club/staircrusher/place/application/port/in/PlaceCacher.kt
@@ -2,7 +2,9 @@ package club.staircrusher.place.application.port.`in`
 
 import club.staircrusher.domain_event.PlaceSearchEvent
 import club.staircrusher.domain_event.dto.PlaceDTO
+import club.staircrusher.place.application.port.out.persistence.BuildingRepository
 import club.staircrusher.place.application.port.out.persistence.PlaceRepository
+import club.staircrusher.place.application.toBuilding
 import club.staircrusher.place.application.toPlace
 import club.staircrusher.stdlib.di.annotation.Component
 import club.staircrusher.stdlib.domain.event.DomainEvent
@@ -13,10 +15,12 @@ import club.staircrusher.stdlib.persistence.TransactionManager
 class PlaceCacher(
     private val transactionManager: TransactionManager,
     private val placeRepository: PlaceRepository,
+    private val buildingRepository: BuildingRepository,
 ) : DomainEventSubscriber<PlaceSearchEvent>() {
     override fun onDomainEvent(event: PlaceSearchEvent) {
         transactionManager.doInTransaction {
             placeRepository.saveAll(event.searchResult.map(PlaceDTO::toPlace))
+            buildingRepository.saveAll(event.searchResult.mapNotNull { it.building?.toBuilding() })
         }
     }
