+++ subprojects/stdlib/src/main/kotlin/club/staircrusher/stdlib/geography/EupMyeonDongs.kt
@@ -1,8 +1,7 @@
 package club.staircrusher.stdlib.geography
 
+import club.staircrusher.stdlib.util.Hashing
 import club.staircrusher.stdlib.util.TextResourceReader
-import java.security.MessageDigest
-import java.util.Base64
 
 
 val eupMyeonDongById = run {
@@ -10,10 +9,10 @@ val eupMyeonDongById = run {
     val eupMyeonDongs = lines.map { line ->
         val (siDo, siGunGu, eupMyeonDong) = line.split("\t")
         EupMyeonDong(
-            id = getHash("$siDo $siGunGu $eupMyeonDong".toByteArray(), length = 36),
+            id = Hashing.getHash("$siDo $siGunGu $eupMyeonDong", length = 36),
             name = eupMyeonDong,
             siGunGu = SiGunGu(
-                id = getHash("$siDo $siGunGu".toByteArray(), length = 36),
+                id = Hashing.getHash("$siDo $siGunGu", length = 36),
                 name = siGunGu,
                 siDo = siDo,
             ),
@@ -21,16 +20,4 @@ val eupMyeonDongById = run {
     }
     require(eupMyeonDongs.size == eupMyeonDongs.map { it.id }.toSet().size)
     eupMyeonDongs.associateBy { it.id }
-}
-
-private fun getHash(byteArray: ByteArray, length: Int? = null): String {
-    val md = MessageDigest.getInstance("SHA-256")
-    md.update(byteArray)
-    return Base64.getEncoder().encodeToString(md.digest()).let {
-        if (length != null) {
-            it.take(length)
-        } else {
-            it
-        }
-    }
 }
